Rename class and files to AsyncWrapper and clean up old duplicates (#38529)
